Move Accordion into SheetSettings component.
