+++ src/features/import/components/Importer/Configure/SheetSettings/index.tsx
@@ -1,5 +1,7 @@
-import { FC } from 'react';
+import { ExpandMore } from '@mui/icons-material';
 import {
+  AccordionDetails,
+  AccordionSummary,
   Box,
   Checkbox,
   FormControl,
@@ -7,12 +9,25 @@ import {
   InputLabel,
   MenuItem,
   Select,
+  styled,
   Typography,
+  useTheme,
 } from '@mui/material';
+import { FC, useState } from 'react';
+import MuiAccordion, { AccordionProps } from '@mui/material/Accordion';
 
 import { ExperimentRow } from '../Mapping';
 import messageIds from 'features/import/l10n/messageIds';
 import { Msg, useMessages } from 'core/i18n';
+
+const Accordion = styled((props: AccordionProps) => (
+  <MuiAccordion disableGutters elevation={0} square {...props} />
+))(() => ({
+  '&:before': {
+    display: 'none',
+  },
+  border: 0,
+}));
 
 export interface ExperimentSheet {
   data: ExperimentRow[];
@@ -36,11 +51,31 @@ const SheetSettings: FC<SheetSettingsProps> = ({
   sheets,
 }) => {
   const messages = useMessages(messageIds);
+  const theme = useTheme();
+  const [settingsExpanded, setSettingsExpanded] = useState(true);
 
   return (
-    <Box>
-      {sheets.length > 1 && (
-        <Box>
+    <Accordion
+      defaultExpanded
+      disableGutters
+      onChange={(ev, isExpanded) => setSettingsExpanded(isExpanded)}
+    >
+      <AccordionSummary
+        expandIcon={<ExpandMore sx={{ color: theme.palette.primary.main }} />}
+      >
+        <Box display="flex" justifyContent="space-between" width="100%">
+          <Typography variant="h5">
+            <Msg id={messageIds.configuration.settings.header} />
+          </Typography>
+          <Typography color={theme.palette.primary.main} paddingRight={1}>
+            {settingsExpanded
+              ? messages.configuration.hide().toLocaleUpperCase()
+              : messages.configuration.show().toLocaleUpperCase()}
+          </Typography>
+        </Box>
+      </AccordionSummary>
+      <AccordionDetails>
+        {sheets.length > 1 && (
           <FormControl size="small">
             <InputLabel>
               <Msg id={messageIds.configuration.settings.sheetSelectLabel} />
@@ -62,18 +97,18 @@ const SheetSettings: FC<SheetSettingsProps> = ({
               <Msg id={messageIds.configuration.settings.sheetSelectHelpText} />
             </FormHelperText>
           </FormControl>
+        )}
+        <Box alignItems="center" display="flex">
+          <Checkbox
+            checked={firstRowIsHeaders}
+            onChange={onChangeFirstRowIsHeaders}
+          />
+          <Typography>
+            <Msg id={messageIds.configuration.settings.firstRowIsHeaders} />
+          </Typography>
         </Box>
-      )}
-      <Box alignItems="center" display="flex">
-        <Checkbox
-          checked={firstRowIsHeaders}
-          onChange={onChangeFirstRowIsHeaders}
-        />
-        <Typography>
-          <Msg id={messageIds.configuration.settings.firstRowIsHeaders} />
-        </Typography>
-      </Box>
-    </Box>
+      </AccordionDetails>
+    </Accordion>
   );
 };
