+++ src/features/import/components/Importer/Configure/index.tsx
@@ -1,30 +1,13 @@
-import {
-  AccordionDetails,
-  AccordionSummary,
-  Box,
-  styled,
-  Typography,
-  useTheme,
-} from '@mui/material';
-import { CompareArrows, ExpandMore } from '@mui/icons-material';
+import { CompareArrows } from '@mui/icons-material';
+import { Box, useTheme } from '@mui/material';
 import { FC, useState } from 'react';
-import MuiAccordion, { AccordionProps } from '@mui/material/Accordion';
 
 import { ExperimentalFieldTypes } from './Mapping/MappingRow';
 import Mapping from './Mapping';
 import messageIds from 'features/import/l10n/messageIds';
+import { useMessages } from 'core/i18n';
 import ZUIEmptyState from 'zui/ZUIEmptyState';
-import { Msg, useMessages } from 'core/i18n';
 import SheetSettings, { ExperimentSheet } from './SheetSettings';
-
-const Accordion = styled((props: AccordionProps) => (
-  <MuiAccordion disableGutters elevation={0} square {...props} />
-))(() => ({
-  '&:before': {
-    display: 'none',
-  },
-  border: 0,
-}));
 
 export interface ConfiguringData {
   columnId: number;
@@ -42,7 +25,6 @@ const Configure: FC<ConfigureProps> = ({ sheets }) => {
   //settings
   const [firstRowIsHeaders, setFirstRowIsHeaders] = useState(true);
   const [selectedSheetId, setSelectedSheetId] = useState(sheets[0].id);
-  const [settingsExpanded, setSettingsExpanded] = useState(true);
 
   //mapping
   const [currentlyConfiguring, setCurrentlyConfiguring] =
@@ -57,43 +39,19 @@ const Configure: FC<ConfigureProps> = ({ sheets }) => {
     <Box display="flex" flexDirection="column" height="100%" overflow="hidden">
       <Box display="flex" flexGrow={1} overflow="hidden">
         <Box display="flex" flexDirection="column" width="50%">
-          <Accordion
-            defaultExpanded
-            disableGutters
-            onChange={(ev, isExpanded) => setSettingsExpanded(isExpanded)}
-          >
-            <AccordionSummary
-              expandIcon={
-                <ExpandMore sx={{ color: theme.palette.primary.main }} />
-              }
-            >
-              <Box display="flex" justifyContent="space-between" width="100%">
-                <Typography variant="h5">
-                  <Msg id={messageIds.configuration.settings.header} />
-                </Typography>
-                <Typography color={theme.palette.primary.main} paddingRight={1}>
-                  {settingsExpanded
-                    ? messages.configuration.hide().toLocaleUpperCase()
-                    : messages.configuration.show().toLocaleUpperCase()}
-                </Typography>
-              </Box>
-            </AccordionSummary>
-            <AccordionDetails>
-              <SheetSettings
-                firstRowIsHeaders={firstRowIsHeaders}
-                onChangeFirstRowIsHeaders={() =>
-                  setFirstRowIsHeaders(!firstRowIsHeaders)
-                }
-                onChangeSelectedSheet={(id: number) => {
-                  setSelectedSheetId(id);
-                  setSelectedColumns([]);
-                  setCurrentlyConfiguring(null);
-                }}
-                selectedSheet={selectedSheetId}
-                sheets={sheets}
-              />
-            </AccordionDetails>
-          </Accordion>
+          <SheetSettings
+            firstRowIsHeaders={firstRowIsHeaders}
+            onChangeFirstRowIsHeaders={() =>
+              setFirstRowIsHeaders(!firstRowIsHeaders)
+            }
+            onChangeSelectedSheet={(id: number) => {
+              setSelectedSheetId(id);
+              setSelectedColumns([]);
+              setCurrentlyConfiguring(null);
+            }}
+            selectedSheet={selectedSheetId}
+            sheets={sheets}
+          />
           <Mapping
             clearCurrentlyConfiguring={() => setCurrentlyConfiguring(null)}
             currentlyConfiguring={currentlyConfiguring}
